refactor(server): restructure server directory layout and move migrations/config/seeders/datasource.js to db/
- updated `.sequelizerc` and adjusted Sequelize config import paths accordingly
