+++ server/models/index.js
@@ -6,7 +6,7 @@ const Sequelize = require("sequelize");
 const process = require("process");
 const basename = path.basename(__filename);
 const env = process.env.NODE_ENV || "development";
-const config = require(__dirname + "/../config/config.cjs")[env];
+const config = require(__dirname + "/../db/config/config.cjs")[env];
 const db = {};
 
 let sequelize;
